# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Financials API",
    "version": "0.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/categories": {
      "get": {
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "key",
                      "name",
                      "isSystem",
                      "color"
                    ],
                    "properties": {
                      "id": {
                        "format": "uuid",
                        "type": "string"
                      },
                      "key": {
                        "type": "string",
                        "nullable": true
                      },
                      "name": {
                        "type": "string"
                      },
                      "isSystem": {
                        "type": "boolean"
                      },
                      "color": {
                        "type": "string",
                        "enum": [
                          "red-400",
                          "orange-400",
                          "amber-400",
                          "yellow-400",
                          "lime-400",
                          "green-400",
                          "emerald-400",
                          "teal-400",
                          "cyan-400",
                          "sky-400",
                          "blue-400",
                          "indigo-400",
                          "violet-400",
                          "purple-400",
                          "fuchsia-400",
                          "pink-400",
                          "rose-400",
                          "slate-400",
                          "gray-400",
                          "zinc-400",
                          "neutral-400",
                          "stone-400"
                        ]
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "color": {
                    "description": "One of: red-400, orange-400, amber-400, yellow-400, lime-400, green-400, emerald-400, teal-400, cyan-400, sky-400, blue-400, indigo-400, violet-400, purple-400, fuchsia-400, pink-400, rose-400, slate-400, gray-400, zinc-400, neutral-400, stone-400",
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "key",
                    "name",
                    "isSystem",
                    "color"
                  ],
                  "properties": {
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "key": {
                      "type": "string",
                      "nullable": true
                    },
                    "name": {
                      "type": "string"
                    },
                    "isSystem": {
                      "type": "boolean"
                    },
                    "color": {
                      "type": "string",
                      "enum": [
                        "red-400",
                        "orange-400",
                        "amber-400",
                        "yellow-400",
                        "lime-400",
                        "green-400",
                        "emerald-400",
                        "teal-400",
                        "cyan-400",
                        "sky-400",
                        "blue-400",
                        "indigo-400",
                        "violet-400",
                        "purple-400",
                        "fuchsia-400",
                        "pink-400",
                        "rose-400",
                        "slate-400",
                        "gray-400",
                        "zinc-400",
                        "neutral-400",
                        "stone-400"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/categories/{id}": {
      "patch": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "color": {
                    "description": "One of: red-400, orange-400, amber-400, yellow-400, lime-400, green-400, emerald-400, teal-400, cyan-400, sky-400, blue-400, indigo-400, violet-400, purple-400, fuchsia-400, pink-400, rose-400, slate-400, gray-400, zinc-400, neutral-400, stone-400",
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "key",
                    "name",
                    "isSystem",
                    "color"
                  ],
                  "properties": {
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "key": {
                      "type": "string",
                      "nullable": true
                    },
                    "name": {
                      "type": "string"
                    },
                    "isSystem": {
                      "type": "boolean"
                    },
                    "color": {
                      "type": "string",
                      "enum": [
                        "red-400",
                        "orange-400",
                        "amber-400",
                        "yellow-400",
                        "lime-400",
                        "green-400",
                        "emerald-400",
                        "teal-400",
                        "cyan-400",
                        "sky-400",
                        "blue-400",
                        "indigo-400",
                        "violet-400",
                        "purple-400",
                        "fuchsia-400",
                        "pink-400",
                        "rose-400",
                        "slate-400",
                        "gray-400",
                        "zinc-400",
                        "neutral-400",
                        "stone-400"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "reassignTo",
            "required": false,
            "description": "Category that receives the rows of the deleted one"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}